# [SPARK-32079][PYTHON] Remove namedtuple hack by replacing built-in pickle to cloudpickle