@HyukjinKwon 2021-11-24
I added a couple of micro benchamrks.