@HyukjinKwon 2021-11-25
Thanks, @viirya.

Merged to master.